remove cruft of ajax stubs from client, for now
moving all api related fetching to a proxy server b/c we’ll need
a custom set of urls for the app anyway. turning off the useFakeData
flags would have rendered real data fetching without jsonp pointless due
to the api’s not using cors

In essence, while we can stub ajax request with ic-ajax, we must stub
only urls to a single server that we have control over. that being the
case, we might as well stub the data returned on the server, where we
can fetch fresh fake data on a regular basis(like vcd does in ruby)